requirements.txt
Babel @ file:///pub/devel/python/python-babel/python-babel-2.12.1-1.noarch/build/dist/Babel-2.12.1-py3-none-any.whl
blinker==1.9.0
chardet @ file:///pub/devel/python/python-chardet/python-chardet-4.0.0-2.noarch/build/dist/chardet-4.0.0-py2.py3-none-any.whl
click==8.1.7
docutils @ file:///pub/devel/python/python-docutils/python-docutils-0.18.1-1.noarch/build/dist/docutils-0.18.1-py2.py3-none-any.whl
Flask==3.1.0
gunicorn==23.0.0
idna @ file:///pub/devel/python/python-idna/python-idna-3.3-1.noarch/build/dist/idna-3.3-py3-none-any.whl
imagesize @ file:///pub/devel/python/python-imagesize/python-imagesize-1.3.0-1.noarch/build/dist/imagesize-1.3.0-py2.py3-none-any.whl
importlib_metadata==8.5.0
iniconfig @ file:///pub/devel/python/python-iniconfig/python-iniconfig-2.0.0-1.noarch/build/dist/iniconfig-2.0.0-py3-none-any.whl
itsdangerous==2.2.0
Jinja2 @ file:///pub/devel/python/python-jinja2/python-jinja2-3.1.2-1.noarch/build/dist/Jinja2-3.1.2-py3-none-any.whl
MarkupSafe @ file:///pub/devel/python/python-markupsafe/python-markupsafe-2.1.2-1.x86_64/build/dist/MarkupSafe-2.1.2-cp39-cp39-cygwin_3_4_6_x86_64.whl
olefile @ file:///pub/devel/python/python-olefile/python-olefile-0.46-3.noarch/build/dist/olefile-0.46-py2.py3-none-any.whl
packaging @ file:///pub/devel/python/python-packaging/python-packaging-21.3-1.noarch/build/dist/packaging-21.3-py3-none-any.whl
Pillow @ file:///home/Hamis/python-imaging/python-imaging-8.4.0-1.x86_64/build/dist/Pillow-8.4.0-cp39-cp39-cygwin_3_3_3_x86_64.whl
platformdirs @ file:///pub/devel/python/python-platformdirs/python-platformdirs-3.1.1-1.noarch/build/dist/platformdirs-3.1.1-py3-none-any.whl
pluggy @ file:///pub/devel/python/python-pluggy/python-pluggy-1.0.0-1.noarch/build/dist/pluggy-1.0.0-py2.py3-none-any.whl
Pygments @ file:///pub/devel/python/python-pygments/python-pygments-2.10.0-1.noarch/build/dist/Pygments-2.10.0-py3-none-any.whl
pyparsing @ file:///pub/devel/python/python-pyparsing/python-pyparsing-3.0.6-1.noarch/build/dist/pyparsing-3.0.6-py3-none-any.whl
pytest @ file:///pub/devel/python/python-pytest/python-pytest-7.3.0-1.noarch/build/dist/pytest-7.3.0-py3-none-any.whl
python-dotenv==1.0.1
requests @ file:///pub/devel/python/python-requests/python-requests-2.27.1-1.noarch/build/dist/requests-2.27.1-py2.py3-none-any.whl
six @ file:///pub/devel/python/python-six/python-six-1.16.0-1.noarch/build/dist/six-1.16.0-py2.py3-none-any.whl
snowballstemmer @ file:///pub/devel/python/python-snowballstemmer/python-snowballstemmer-2.2.0-1.noarch/build/dist/snowballstemmer-2.2.0-py2.py3-none-any.whl
Sphinx @ file:///pub/devel/python/python-sphinx/python-sphinx-4.4.0-1.noarch/build/dist/Sphinx-4.4.0-py3-none-any.whl
sphinxcontrib-serializinghtml @ file:///pub/devel/python/python-sphinxcontrib-serializinghtml/python-sphinxcontrib-serializinghtml-1.1.5-1.noarch/build/dist/sphinxcontrib_serializinghtml-1.1.5-py2.py3-none-any.whl
toml @ file:///pub/devel/python/python-toml/python-toml-0.10.2-2.noarch/build/dist/toml-0.10.2-py2.py3-none-any.whl
urllib3 @ file:///pub/devel/python/python-urllib3/python-urllib3-1.26.7-1.noarch/build/dist/urllib3-1.26.7-py2.py3-none-any.whl
Werkzeug==3.1.3
zipp==3.21.0
